+++ src/main/java/study/racing/model/car/Car.java
@@ -1,4 +1,4 @@
-package study.car.race;
+package study.racing.model.car;
 
 public class Car {
