+++ src/main/java/study/racing/utils/RandomUtils.java
@@ -1,4 +1,4 @@
-package study.car.race;
+package study.racing.utils;
 
 import java.util.Random;
