+++ src/test/java/study/racing/CarTest.java
@@ -1,10 +1,12 @@
-package study.car.race;
+package study.racing;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+
+import study.racing.model.car.Car;
 
 public class CarTest {
